+++ SmartReceipts/WBTripsViewController.m
@@ -144,7 +144,7 @@ static NSString *CellIdentifier = @"Cell";
     
     UILabel *priceField = _protoCell.priceField;
     
-    for (int i = 0; i < [_trips count]; ++i) {
+    for (NSUInteger i = 0; i < [_trips count]; ++i) {
         NSString *str = [[_trips tripAtIndex:i] priceWithCurrencyFormatted];
         
         // dumb, but works better than calculating bounds for attributed text because dynamically includes paddings etc.
